requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
python-dotenv==1.0.1
google-generativeai==0.8.3
pydantic==2.10.5
typing-extensions==4.12.2
httpx==0.28.1
websockets==13.1 